app/requirements.txt
asgiref==3.2.10
autopep8==1.5.4
Django==3.1
django-crispy-forms
gunicorn==20.0.4
pycodestyle==2.6.0
pytz==2020.1
sqlparse==0.3.1
toml==0.10.1
whitenoise
dj-database-url
djangorestframework
django-cors-headers
django-rest-swagger
djangorestframework-simplejwt
drf-yasg
django-debug-toolbar
numpy==1.19.4
pandas==1.1.5
xlrd==1.2.0
requests
celery
redis
